separate summary page & implement editMode
